Move canister trait into a separate file
